Adding a debug log for the confirmation token
In development if you need the confirmation token, you can only reach it
through the email, which more often than not is stored in memory.

This adds a simple way to use the code and since it's only on debug log level
it should never affect anything on production environments. Even if the production
log level is debug, the information is not as sensitive as a password